fix: rename mapping sync and synced
