+++ test/util/rest/data/synced.spec.js
@@ -42,7 +42,7 @@ describe('Testing synced', { useTmpDir: true }, () => {
   });
 
   it('Test documents are synced', async ({ dir }) => {
-    expect(await index.rest.mapping.sync('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
+    expect(await index.rest.mapping.apply('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
     expect(await index.rest.data.update('offer', [{
       action: 'update',
       doc: index.data.remap('offer', {
@@ -54,7 +54,7 @@ describe('Testing synced', { useTmpDir: true }, () => {
     index.model.register('offer', updatedOfferModel);
     index.index.register('offer', updatedOfferIndex);
     await createIndexVersion(dir);
-    expect(await index.rest.mapping.sync('offer')).to.deep.equal(['offer@e35ec51a3c35e2d9982e1ac2bbe23957a637a9e0']);
+    expect(await index.rest.mapping.apply('offer')).to.deep.equal(['offer@e35ec51a3c35e2d9982e1ac2bbe23957a637a9e0']);
     expect(await index.rest.data.update('offer', [{
       action: 'update',
       doc: index.data.remap('offer', {
@@ -68,7 +68,7 @@ describe('Testing synced', { useTmpDir: true }, () => {
   });
 
   it('Test documents are not synced', async ({ dir }) => {
-    expect(await index.rest.mapping.sync('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
+    expect(await index.rest.mapping.apply('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
     expect(await index.rest.data.update('offer', [{
       action: 'update',
       doc: index.data.remap('offer', {
@@ -81,12 +81,12 @@ describe('Testing synced', { useTmpDir: true }, () => {
     index.model.register('offer', updatedOfferModel);
     index.index.register('offer', updatedOfferIndex);
     await createIndexVersion(dir);
-    expect(await index.rest.mapping.sync('offer')).to.deep.equal(['offer@e35ec51a3c35e2d9982e1ac2bbe23957a637a9e0']);
+    expect(await index.rest.mapping.apply('offer')).to.deep.equal(['offer@e35ec51a3c35e2d9982e1ac2bbe23957a637a9e0']);
     expect(await index.rest.data.synced('offer')).to.equal(false);
   });
 
   it('Test version does not exist', async ({ dir }) => {
-    expect(await index.rest.mapping.sync('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
+    expect(await index.rest.mapping.apply('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
     instantiateIndex();
     index.model.register('offer', updatedOfferModel);
     index.index.register('offer', updatedOfferIndex);
