+++ test/util/rest/mapping/applied.spec.js
@@ -3,7 +3,7 @@ const { describe } = require('node-tdd');
 const Index = require('../../../../src/index');
 const { registerEntitiesForIndex } = require('../../../helper');
 
-describe('Testing synced', {
+describe('Testing applied', {
   useTmpDir: true
 }, () => {
   let index;
@@ -17,16 +17,16 @@ describe('Testing synced', {
     expect(await index.rest.mapping.delete('offer')).to.equal(true);
   });
 
-  it('Test mapping is synced', async ({ dir }) => {
+  it('Test mapping is applied', async ({ dir }) => {
     expect(index.index.versions.persist(dir)).to.equal(true);
     index.index.versions.load(dir);
-    expect(await index.rest.mapping.sync('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
-    expect(await index.rest.mapping.synced('offer')).to.equal(true);
+    expect(await index.rest.mapping.apply('offer')).to.deep.equal(['offer@6a1b8f491e156e356ab57e8df046b9f449acb440']);
+    expect(await index.rest.mapping.applied('offer')).to.equal(true);
   });
 
-  it('Test mapping is not synced', async ({ dir }) => {
+  it('Test mapping is not applied', async ({ dir }) => {
     expect(index.index.versions.persist(dir)).to.equal(true);
     index.index.versions.load(dir);
-    expect(await index.rest.mapping.synced('offer')).to.equal(false);
+    expect(await index.rest.mapping.applied('offer')).to.equal(false);
   });
 });
